+++ src/main/java/project/domain/dataAccess/CultureRepository.java
@@ -3,6 +3,7 @@ package project.domain.dataAccess;
 import oracle.jdbc.OracleTypes;
 import project.ui.console.utils.Utils;
 
+import java.math.BigDecimal;
 import java.sql.CallableStatement;
 import java.sql.Connection;
 import java.sql.ResultSet;
@@ -10,10 +11,10 @@ import java.sql.SQLException;
 import java.util.*;
 
 public class CultureRepository {
-    public Map<Integer, String> getCultures() throws SQLException {
+    public Map<BigDecimal, String> getCultures() throws SQLException {
         CallableStatement callStmt = null;
         ResultSet resultSet = null;
-        Map<Integer, String> map = new HashMap<>();
+        Map<BigDecimal, String> map = new HashMap<>();
 
         try {
             Connection connection = DatabaseConnection.getInstance().getConnection();
